+++ src/AZERTYGlobal.Tests/GlobalUsings.cs
@@ -1,1 +1,3 @@
 global using TypingEngine.Core;
+global using TypingEngine.Windows;
+global using TypingEngine.Windows.Testing;