+++ src/TypingEngine.Windows.Tests/GlobalUsings.cs
@@ -1,2 +1,4 @@
 global using TypingEngine.Core;
 global using TypingEngine.Windows;
+global using TypingEngine.Windows.Testing;
+global using Xunit;